Deleted old kbn-top-nav directive (#43168)
* Deleted old kbn-top-nav directive.
Can be merged only when monitoring replace their last usage.

* fix texts

* Remove CSS import
Removed fix to config in monitoring (fix is on #43187)

* rename monitoring directive
