+++ src/legacy/core_plugins/console/public/index.html
@@ -1,7 +1,7 @@
-<kbn-top-nav-v2 
+<kbn-top-nav 
   app-name="'console'"
   config="topNavMenu" 
-></kbn-top-nav-v2>
+></kbn-top-nav>
 <kbn-dev-tools-app data-test-subj="console">
   <sense-history ng-show="showHistory" is-shown="showHistory" close="closeHistory()" history-dirty="lastRequestTimestamp"></sense-history>
   <div class="conApp">
