+++ src/legacy/core_plugins/kibana/public/dashboard/dashboard_app.html
@@ -3,7 +3,7 @@
   ng-class="{'dshAppContainer--withMargins': model.useMargins}"
 >
   <!-- Local nav. -->
-  <kbn-top-nav-v2 
+  <kbn-top-nav 
     ng-show="chrome.getVisible()"
     app-name="'dashboard'"
     config="topNavMenu" 
@@ -23,7 +23,7 @@
     is-refresh-paused="model.refreshInterval.pause"
     refresh-interval="model.refreshInterval.value"
     on-refresh-change="onRefreshChange">
-  </kbn-top-nav-v2>
+  </kbn-top-nav>
 
   <!--
   The top nav is hidden in embed mode but the filter bar must still be present so
