+++ src/legacy/core_plugins/kibana/public/discover/index.html
@@ -1,7 +1,7 @@
 <discover-app class="app-container">
   <!-- Local nav. -->
 
-  <kbn-top-nav-v2
+  <kbn-top-nav
     app-name="'discover'"
     config="topNavMenu"
     show-search-bar="true"
@@ -18,7 +18,7 @@
     refresh-interval="refreshInterval.value"
     on-refresh-change="onRefreshChange"
   >
-  </kbn-top-nav-v2>
+  </kbn-top-nav>
 
   <main class="container-fluid">
     <div class="row">
