+++ src/legacy/core_plugins/kibana/public/visualize/editor/editor.html
@@ -32,7 +32,7 @@
     All visualizaions also have least a timepicker \ autorefresh component, which is why
     show-query-bar is set to "true".
   -->
-  <kbn-top-nav-v2 
+  <kbn-top-nav 
     app-name="'visualize'"
     config="topNavMenu"
     show-search-bar="true"
@@ -53,7 +53,7 @@
     refresh-interval="refreshInterval.value"
     on-refresh-change="onRefreshChange"
   >
-  </kbn-top-nav-v2>
+  </kbn-top-nav>
 
   <!--
     The top nav is hidden in embed mode but the filter bar must still be present so
