+++ src/legacy/core_plugins/timelion/public/index.html
@@ -11,7 +11,7 @@
     </span>
   </span>
   <!-- Local nav. -->
-  <kbn-top-nav-v2 
+  <kbn-top-nav 
     app-name="'timelion'"
     config="topNavMenu"
     show-search-bar="true"
@@ -24,7 +24,7 @@
     refresh-interval="model.refreshInterval.value"
     on-refresh-change="onRefreshChange"
     on-query-submit="onTimeUpdate">
-  </kbn-top-nav-v2>
+  </kbn-top-nav>
 
 
   <div class="timApp__menus">
