+++ src/legacy/ui/public/_index.scss
@@ -18,7 +18,6 @@
 @import './error_url_overflow/index';
 @import './exit_full_screen/index';
 @import './field_editor/index';
-@import './kbn_top_nav/index';
 @import './notify/index';
 @import './share/index';
 @import './style_compile/index';
