--- src/legacy/ui/public/kbn_top_nav/_index.scss
@@ -1,1 +1,0 @@
-@import './kbn_top_nav';
