--- src/legacy/ui/public/kbn_top_nav/_kbn_top_nav.scss
@@ -1,36 +1,0 @@
-/**
- * 1. Make sure the timepicker is always one right, even if the main menu doesn't exist
- */
-
-kbn-top-nav {
-  z-index: 5;
-}
-
-.kbnTopNav {
-  background-color: $euiPageBackgroundColor;
-  border-bottom: $euiBorderThin;
-}
-
-.kbnTopNav__row {
-  display: flex;
-  align-items: center;
-  justify-content: flex-end; /* 1 */
-}
-
-.kbnTopNav__mainMenu {
-  flex-grow: 1; /* 1 */
-
-  &:empty {
-    // Make sure the area doesn't take up any space if it is empty
-    padding: 0;
-  }
-
-  .kuiLocalMenuItem {
-    @include euiFontSizeS;
-    font-weight: $euiFontWeightMedium;
-  }
-
-  .kuiLocalMenuItem:not(.kuiLocalMenuItem-isDisabled) {
-    color: $euiLinkColor;
-  }
-}
