--- src/legacy/ui/public/kbn_top_nav/kbn_top_nav.html
@@ -1,47 +1,0 @@
-<div class="kbnTopNav" ng-show="kbnTopNav.isVisible()" data-test-subj="top-nav">
-  <!-- Top row slot, e.g. breadcrumbs, title. -->
-  <div data-transclude-slot="topLeftCorner"></div>
-
-  <div class="kbnTopNav__row">
-    <!-- Right-side: menu -->
-
-    <div class="kuiLocalMenu kbnTopNav__mainMenu" ng-if="kbnTopNav.menuItems.length">
-      <button
-        class="kuiLocalMenuItem kbn-resetFocusState {{menuItem.className}}"
-        ng-repeat="menuItem in kbnTopNav.menuItems"
-        aria-label="{{::menuItem.description}}"
-        aria-haspopup="{{!menuItem.hasFunction}}"
-        aria-expanded="{{kbnTopNav.isCurrent(menuItem.key)}}"
-        aria-disabled="{{menuItem.disableButton()}}"
-        ng-class="{'kuiLocalMenuItem-isSelected': kbnTopNav.isCurrent(menuItem.key), 'kuiLocalMenuItem-isDisabled': menuItem.disableButton()}"
-        ng-click="kbnTopNav.handleClick(menuItem, $event)"
-        ng-bind="menuItem.label"
-        tooltip="{{menuItem.tooltip()}}"
-        tooltip-placement="bottom"
-        tooltip-popup-delay="400"
-        tooltip-append-to-body="1"
-        data-test-subj="{{menuItem.testId}}"
-      ></button>
-    </div>
-
-    <!-- Time-picker "menu item" -->
-    <kbn-global-timepicker></kbn-global-timepicker>
-  </div>
-
-  <!-- Dropdown content, e.g. time-picker. -->
-  <div class="kuiLocalDropdown" ng-show="kbnTopNav.rendered">
-    <button
-      class="kuiLocalDropdownCloseButton"
-      ng-click="kbnTopNav.close()"
-      aria-label="{{ ::'common.ui.topNav.closeAriaLabel' | i18n: { defaultMessage: 'Close' } }}"
-    >
-      <span class="kuiIcon fa-chevron-circle-up"></span>
-    </button>
-    <div id="template_wrapper"><!-- Content gets dynamically inserted here. --></div>
-  </div>
-
-  <div class="kuiLocalNavRow kuiLocalNavRow--secondary">
-    <!-- Bottom row slot, e.g. tabs. -->
-    <div data-transclude-slot="bottomRow"></div>
-  </div>
-</div>
