+++ src/legacy/ui/public/kbn_top_nav/kbn_top_nav.js
@@ -17,158 +17,109 @@
  * under the License.
  */
 
-/**
- * A configuration object for a top nav component.
- * @typedef {Object} KbnTopNavConfig
- * @type Object
- * @property {string} key - identifier of menu item.
- * @property {string} label - A display string which will be shown in the top nav for this option.
- * @property {string} [description] - optional, used for the screen-reader description of this
- *  menu. Defaults to "Toggle ${label} view" for templated menu items and just "${label}" for
- *  programmatic menu items
- * @property {string} testId - for testing purposes, can be used to retrieve this item.
- * @property {Object} [template] - an html template that will be shown when this item is clicked.
- *  If template is not given then run should be supplied.
- * @property {function} [run] - an optional function that will be run when the nav item is clicked.
- *  Either this or template parameter should be specified.
- * @param {boolean} [hideButton] - optional, set to true to prevent a menu item from being created.
- *  This allow injecting templates into the navbar that don't have an associated template
- */
+import 'ngreact';
+import { wrapInI18nContext } from 'ui/i18n';
+import { uiModules } from 'ui/modules';
+import { TopNavMenu } from '../../../core_plugins/kibana_react/public';
+import { Storage } from 'ui/storage';
+import chrome from 'ui/chrome';
 
-/**
- * kbnTopNav directive
- *
- * The top section that optionally shows the timepicker and menu items.
- *
- * ```
- * <kbn-top-nav name="current-app-for-extensions" config="path.to.menuItems" ></kbn-top-nav>
- * ```
- *
- * Menu items/templates are passed to the kbnTopNav via the config attribute
- * and should be defined as an array of objects. Each object represents a menu
- * item and should be of type kbnTopNavConfig.
- *
- * @param {Array<kbnTopNavConfig>|KbnTopNavController} config
- *
- * Programmatic control of the navbar can be achieved one of two ways
- */
-
-import _ from 'lodash';
-import angular from 'angular';
-import './timepicker';
-import '../directives/watch_multi';
-import '../directives/input_focus';
-import { uiModules } from '../modules';
-import template from './kbn_top_nav.html';
-import { KbnTopNavControllerProvider } from './kbn_top_nav_controller';
-import { NavBarExtensionsRegistryProvider } from '../registry/navbar_extensions';
 
 const module = uiModules.get('kibana');
 
-module.directive('kbnTopNav', function (Private) {
-  const KbnTopNavController = Private(KbnTopNavControllerProvider);
-  const navbarExtensions = Private(NavBarExtensionsRegistryProvider);
-  const getNavbarExtensions = _.memoize(function (name) {
-    if (!name) throw new Error('navbar directive requires a name attribute');
-    return _.sortBy(navbarExtensions.byAppName[name], 'order');
-  });
-
+module.directive('kbnTopNav', () => {
   return {
     restrict: 'E',
-    transclude: true,
-    template,
+    template: '',
+    compile: (elem) => {
+      const child = document.createElement('kbn-top-nav-helper');
 
-    // TODO: The kbnTopNav currently requires that it share a scope with
-    // its parent directive. This allows it to export the kbnTopNav controller
-    // and allows the config templates to use values from the parent scope.
-    //
-    // Moving this to an isolate scope will require modifying the config
-    // directive to support child directives, instead of templates, so that
-    // parent controllers can be imported/required rather than simply referenced
-    // directly in the template.
-    //
-    // TODO: Our fake multi-slot transclusion solution also depends on an inherited
-    // scope. Moving this to an isolate scope will cause this to break.
-    //
-    // scope: {}
-
-    controller($scope, $attrs, $element, $transclude) {
-      // This is a semi-hacky solution to missing slot-transclusion support in Angular 1.4.7
-      // (it was added as a core feature in 1.5). Borrowed from http://stackoverflow.com/a/22080765.
-      $scope.transcludes = {};
-
-      // Extract transcluded elements for use in the link function.
-      $transclude(clone => {
-        // We expect the transcluded elements to be wrapped in a single div.
-        const transcludedContentContainer = _.find(clone, item => {
-          if (item.attributes) {
-            return _.find(item.attributes, attr => {
-              return attr.name.indexOf('data-transclude-slots') !== -1;
-            });
-          }
-        });
-
-        if (!transcludedContentContainer) {
-          return;
-        }
-
-        const transcludedContent = transcludedContentContainer.children;
-        _.forEach(transcludedContent, transcludedItem => {
-          const transclusionSlot = transcludedItem.getAttribute('data-transclude-slot');
-          $scope.transcludes[transclusionSlot] = transcludedItem;
-        });
-      });
-      const extensions = getNavbarExtensions($attrs.name);
-
-      function initTopNav(newConfig, oldConfig) {
-        if (_.isEqual(oldConfig, newConfig)) return;
-
-        if (newConfig instanceof KbnTopNavController) {
-          newConfig.addItems(extensions);
-          $scope.kbnTopNav = new KbnTopNavController(newConfig);
-        } else {
-          newConfig = newConfig.concat(extensions);
-          $scope.kbnTopNav = new KbnTopNavController(newConfig);
-        }
-        $scope.kbnTopNav._link($scope, $element);
+      // Copy attributes to the child directive
+      for (const attr of elem[0].attributes) {
+        child.setAttribute(attr.name, attr.value);
       }
 
-      const getTopNavConfig = () => {
-        return _.get($scope, $attrs.config, []);
+      // Add a special attribute that will change every time that one
+      // of the config array's disableButton function return value changes.
+      child.setAttribute('disabled-buttons', 'disabledButtons');
+
+      // Pass in storage
+      const localStorage = new Storage(window.localStorage);
+      child.setAttribute('store', 'store');
+      child.setAttribute('ui-settings', 'uiSettings');
+
+      // Append helper directive
+      elem.append(child);
+
+      const linkFn = ($scope, _, $attr) => {
+        $scope.store = localStorage;
+        $scope.uiSettings = chrome.getUiSettingsClient();
+
+        // Watch config changes
+        $scope.$watch(() => {
+          const config = $scope.$eval($attr.config) || [];
+          return config.map((item) => {
+            // Copy key into id, as it's a reserved react propery.
+            // This is done for Angular directive backward compatibility.
+            // In React only id is recognized.
+            if (item.key && !item.id) {
+              item.id = item.key;
+            }
+
+            // Watch the disableButton functions
+            if (typeof item.disableButton === 'function') {
+              return item.disableButton();
+            }
+            return item.disableButton;
+          });
+        }, (newVal) => {
+          $scope.disabledButtons = newVal;
+        },
+        true);
       };
 
-      const topNavConfig = getTopNavConfig();
-
-      // Because we store $scope and $element on the kbnTopNavController, if this was passed an instance
-      // instead of a configuration, it will enter an infinite digest loop. Only watch for updates if a config
-      // was passed instead. This is ugly, but without diving into a larger refactor, the smallest temporary solution
-      // to get dynamic nav updates working for dashboard. Console is currently the only place that passes a
-      // KbnTopNavController (and a slew of tests).
-      if (!(topNavConfig instanceof KbnTopNavController)) {
-        $scope.$watch(getTopNavConfig, initTopNav, true);
-      }
-
-      initTopNav(topNavConfig, null);
-
-      if (!_.has($scope, 'showTimepickerInTopNav')) {
-        $scope.showTimepickerInTopNav = true;
-      }
-
-      return $scope.kbnTopNav;
-    },
-
-    link(scope) {
-      // These are the slots where transcluded elements can go.
-      const transclusionSlotNames = ['topLeftCorner', 'bottomRow'];
-
-      // Transclude elements into specified "slots" in the top nav.
-      transclusionSlotNames.forEach(name => {
-        const transcludedItem = scope.transcludes[name];
-        if (transcludedItem) {
-          const transclusionSlot = document.querySelector(`[data-transclude-slot="${name}"]`);
-          angular.element(transclusionSlot).replaceWith(transcludedItem);
-        }
-      });
+      return linkFn;
     }
   };
+});
+
+module.directive('kbnTopNavHelper', (reactDirective) => {
+  return reactDirective(
+    wrapInI18nContext(TopNavMenu),
+    [
+      ['name', { watchDepth: 'reference' }],
+      ['config', { watchDepth: 'value' }],
+      ['disabledButtons', { watchDepth: 'reference' }],
+
+      ['query', { watchDepth: 'reference' }],
+      ['store', { watchDepth: 'reference' }],
+      ['uiSettings', { watchDepth: 'reference' }],
+      ['intl', { watchDepth: 'reference' }],
+      ['store', { watchDepth: 'reference' }],
+
+      ['onQuerySubmit', { watchDepth: 'reference' }],
+      ['onFiltersUpdated', { watchDepth: 'reference' }],
+      ['onRefreshChange', { watchDepth: 'reference' }],
+
+      ['indexPatterns', { watchDepth: 'collection' }],
+      ['filters', { watchDepth: 'collection' }],
+
+      // All modifiers default to true.
+      // Set to false to hide subcomponents.
+      'showSearchBar',
+      'showFilterBar',
+      'showQueryBar',
+      'showQueryInput',
+      'showDatePicker',
+
+      'appName',
+      'screenTitle',
+      'dateRangeFrom',
+      'dateRangeTo',
+      'isRefreshPaused',
+      'refreshInterval',
+      'disableAutoFocus',
+      'showAutoRefreshOnly',
+    ],
+  );
 });
