--- src/legacy/ui/public/kbn_top_nav/timepicker/kbn_global_timepicker.html
@@ -1,20 +1,0 @@
-<div
-  ng-if="timefilterValues.isAutoRefreshSelectorEnabled || timefilterValues.isTimeRangeSelectorEnabled"
-  data-shared-timefilter-from="{{ getSharedTimeFilterFromDate() }}"
-  data-shared-timefilter-to="{{ getSharedTimeFilterToDate() }}"
-  class="kuiLocalMenu"
-  data-test-subj="globalTimepicker"
->
-  <super-date-picker
-    start="timefilterValues.time.from"
-    end="timefilterValues.time.to"
-    is-paused="timefilterValues.refreshInterval.pause"
-    refresh-interval="timefilterValues.refreshInterval.value"
-    on-time-change="updateFilter"
-    on-refresh-change="updateInterval"
-    is-auto-refresh-only="!timefilterValues.isTimeRangeSelectorEnabled"
-    commonly-used-ranges="commonlyUsedRanges"
-    date-format="dateFormat"
-    recently-used-ranges="recentlyUsedRanges"
-  />
-</div>
