+++ x-pack/legacy/plugins/graph/public/templates/index.html
@@ -1,7 +1,7 @@
 <div id="graphBasic" ng-controller="graphuiPlugin">
   <!-- Local nav. -->
-  <kbn-top-nav-v2 name="workspacesTopNav" config="topNavMenu">
-  </kbn-top-nav-v2>
+  <kbn-top-nav name="workspacesTopNav" config="topNavMenu">
+  </kbn-top-nav>
 
   <div class="gphGraph__menus" ng-show="menus.showLoad || menus.showSave || menus.showSettings">
     <graph-load ng-show="menus.showLoad"></graph-load>
