+++ x-pack/legacy/plugins/reporting/export_types/common/layouts/preserve_layout.css
@@ -15,7 +15,6 @@
 }
 
 /* hide unusable controls */
-.kbnGlobalNav,
 kbn-top-nav,
 filter-bar,
 ::-webkit-scrollbar,
