+++ x-pack/legacy/plugins/reporting/export_types/common/layouts/print.css
@@ -15,7 +15,6 @@
 }
 
 /* hide unusable controls */
-.kbnGlobalNav,
 kbn-top-nav,
 filter-bar,
 ::-webkit-scrollbar,
