+++ x-pack/plugins/translations/translations/ja-JP.json
@@ -554,8 +554,6 @@
     "common.ui.stateManagement.unableToParseUrlErrorMessage": "URL をパースできません",
     "common.ui.stateManagement.unableToRestoreUrlErrorMessage": "URL を完全に復元できません。共有機能を使用していることを確認してください。",
     "common.ui.stateManagement.unableToStoreHistoryInSessionErrorMessage": "セッションがいっぱいで安全に削除できるアイテムが見つからないため、Kibana は履歴アイテムを保存できません。\n\nこれは大抵新規タブに移動することで解決されますが、より大きな問題が原因である可能性もあります。このメッセージが定期的に表示される場合は、{gitHubIssuesUrl} で問題を報告してください。",
-    "common.ui.topNav.closeAriaLabel": "閉じる",
-    "common.ui.topNav.toggleViewAriaLabel": "{optLabel} ビューを切り替える",
     "common.ui.url.replacementFailedErrorMessage": "置換に失敗、未解決の表現式: {expr}",
     "common.ui.url.savedObjectIsMissingNotificationMessage": "保存されたオブジェクトがありません",
     "common.ui.vis.aggConfig.percentageOfLabel": "{label} のパーセンテージ",
