+++ x-pack/plugins/translations/translations/zh-CN.json
@@ -554,8 +554,6 @@
     "common.ui.stateManagement.unableToParseUrlErrorMessage": "无法解析 URL",
     "common.ui.stateManagement.unableToRestoreUrlErrorMessage": "无法完整还原 URL，确保使用共享功能。",
     "common.ui.stateManagement.unableToStoreHistoryInSessionErrorMessage": "Kibana 无法将历史记录项存储在您的会话中，因为其已满，并且似乎没有任何可安全删除的项。\n\n通常可通过移至新的标签页来解决此问题，但这会导致更大的问题。如果您有规律地看到此消息，请在 {gitHubIssuesUrl} 提交问题。",
-    "common.ui.topNav.closeAriaLabel": "关闭",
-    "common.ui.topNav.toggleViewAriaLabel": "切换 {optLabel} 视图",
     "common.ui.url.replacementFailedErrorMessage": "替换失败，未解析的表达式：{expr}",
     "common.ui.url.savedObjectIsMissingNotificationMessage": "已保存对象缺失",
     "common.ui.vis.aggConfig.percentageOfLabel": "{label} 的百分比",
